fix randomness
Three elves on the same page ended up with identical stats.  With these changes, they now have totally different stats, even when placed on the same page.

Randomness is no longer so dependent upon the page or chapter, but now defaults to a number, which cycles round once per creature.