big projectiles edit
Rolling a tie with a missile resulted in 'hitting the person behind the
initial target, if anyone', which won't always feel satisfying.  The new
rule is simply that the attacker fumbles the missile, so they waste an
AP but keep the missile so they can try again.

The structure's now cleanly divided into general projectile rules and
projectile weapons.